Resume Start.
Not very much was added in. To be exact, my name as a title, a "background" image and a nav tab is all that is currently in there along with some CSS to make it look a bit better. 
I actually do plan to work on this term project at least several instances in my own time, so there could be more changes to come pretty quickly.