Good and Bad News
Bad news is that I still am super slow at getting progress on this, even with five days off. I just can't seem to find the motivation to do so and when I do, I'm not very good at troubleshooting myself and I keep running into issues that stump me to lose confidence.

For the good news, I now have a game plan. This week, I'll split my work by working on a section per day as there are five of them. For the first time in a while, I feel like that will give me a good routine to work with along with a better work ethic behind it. I''ll be very mad at myself if I continue to procrastinate with this even further, but I'll do my best this time for sure because this is something I know I can own up to with each day not having that much to be done comparatively.