I did it.
It may have been a day late, but I can safety say that I am satisfied with my work here today, which I cannot say for all. I however can improve even more on it, and will chose to do so, even past the grading process.

I'm glad this was finally accomplished, and I hope you enjoy it in this current form and beyond.