I did a lot of things now.
Funny enough, now is when I find a great deal of motivation in working on this passion of mine.

The three biggest changes visually are all text lines being centered on the page, the links being further separated by that respective class and the backgrounds made even darker for added visibility on the text.

There are also some other niffy changes I've done for further polish and all documentation along with it shall continue for a while as I now cherish this site of  mine.